Backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.0.3
psutil==7.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2023.8.8
redis==4.5.5
SQLAlchemy==1.4.41